+++ client/src/lib/lenis.ts
@@ -1,0 +1,19 @@
+import Lenis from 'lenis'
+import { gsap } from './gsap'
+
+export function initLenis(): Lenis {
+  const lenis = new Lenis({
+    duration: 1.2,
+    easing: (t) => Math.min(1, 1.001 - Math.pow(2, -10 * t)),
+    orientation: 'vertical',
+    smoothWheel: true,
+  })
+
+  gsap.ticker.add((time) => {
+    lenis.raf(time * 1000)
+  })
+
+  gsap.ticker.lagSmoothing(0)
+
+  return lenis
+}